+++ SEPADirectDebit/src/main/java/com/braintreepayments/api/SEPADirectDebitLauncherCallback.java
@@ -5,9 +5,9 @@ import androidx.annotation.NonNull;
 /**
  * Used to receive notification that the SEPA mandate web browser flow completed.
  * Once this is invoked, continue the flow by calling
- * {@link SEPADirectDebitClient#onBrowserSwitchResult(SEPADirectDebitBrowserSwitchResult, SEPADirectDebitBrowserSwitchResultCallback)}
+ * {@link SEPADirectDebitClient#tokenize(SEPADirectDebitPaymentAuthResult, SEPADirectDebitBrowserSwitchResultCallback)}
  */
 public interface SEPADirectDebitLauncherCallback {
 
-    void onResult(@NonNull SEPADirectDebitBrowserSwitchResult sepaDirectDebitBrowserSwitchResult);
+    void onResult(@NonNull SEPADirectDebitPaymentAuthResult paymentAuthResult);
 }