+++ SEPADirectDebit/src/main/java/com/braintreepayments/api/SEPADirectDebitPaymentAuthResult.java
@@ -3,19 +3,19 @@ package com.braintreepayments.api;
 /**
  * Result received from the SEPA mandate web flow through {@link SEPADirectDebitBrowserSwitchResultCallback}.
  * This result should be passed to
- * {@link SEPADirectDebitClient#onBrowserSwitchResult(SEPADirectDebitBrowserSwitchResult, SEPADirectDebitBrowserSwitchResultCallback)} )}
+ * {@link SEPADirectDebitClient#tokenize(SEPADirectDebitPaymentAuthResult, SEPADirectDebitBrowserSwitchResultCallback)} )}
  * to complete the SEPA mandate flow.
  */
-public class SEPADirectDebitBrowserSwitchResult {
+public class SEPADirectDebitPaymentAuthResult {
 
     private BrowserSwitchResult browserSwitchResult;
     private Exception error;
 
-    SEPADirectDebitBrowserSwitchResult(BrowserSwitchResult browserSwitchResult) {
+    SEPADirectDebitPaymentAuthResult(BrowserSwitchResult browserSwitchResult) {
         this.browserSwitchResult = browserSwitchResult;
     }
 
-    SEPADirectDebitBrowserSwitchResult(Exception error) {
+    SEPADirectDebitPaymentAuthResult(Exception error) {
         this.error = error;
     }